doc: classify Chrome DevTools Protocol as tier 2
Chrome DevTools Protocol is the de-facto debugger protocol for
Node.js and is tested regularly in Node.js test suite / CI. It matches
the tier 2 classification but lacks proper documentation in Node.js.

Classify the Chrome DevTools Protocol as tier 2 with a target of
tier 1 once we have proper documentation on stable protocols.

On the other hand, the Chrome DevTools Frontend is not shipped with
Node.js and not tested in Node.js CI. So keep it as tier 3.

PR-URL: https://github.com/nodejs/node/pull/57634
Reviewed-By: Yagiz Nizipli <yagiz@nizipli.com>
Reviewed-By: Marco Ippolito <marcoippolito54@gmail.com>
Reviewed-By: Matteo Collina <matteo.collina@gmail.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Darshan Sen <raisinten@gmail.com>
Reviewed-By: Rafael Gonzaga <rafael.nunu@hotmail.com>
Reviewed-By: Moshe Atlow <moshe@atlow.co.il>